improved validations in server

+++ 21Lane/server.py
@@ -8,15 +8,15 @@ from pyftpdlib.authorizers import DummyAuthorizer
 import os 
 import socket
 from multiprocessing import Process 
+from os.path import exists as path_exists
+from customErrors import PortUnavailableError
+
 
 def isPortAvailable(port):
     sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
     result = sock.connect_ex(('127.0.0.1',port))
     return not (result == 0)
 
-
-class PortUnavailable(Exception):
-    pass 
 
 
 class CustomHandler(FTPHandler):
@@ -40,6 +40,7 @@ class Server:
     def __init__(self):
         self.port = 2121
         self.sharedDir = ""
+        self.is_running = False 
         self.dtp_handler = ThrottledDTPHandler
         self.ftp_handler = CustomHandler
         self.ftp_handler.banner = "21Lane ready"
@@ -48,9 +49,11 @@ class Server:
         if isPortAvailable(port):
             self.port = port    
         else:
-            raise PortUnavailable
+            raise PortUnavailableError
 
     def setSharedDirectory(self, path):
+        if not path_exists(path):
+            raise FileNotFoundError
         self.authorizer = DummyAuthorizer()
         self.authorizer.add_anonymous(path)
         self.ftp_handler.authorizer = self.authorizer
@@ -63,9 +66,14 @@ class Server:
         self.server = FTPServer(('', self.port), self.ftp_handler)
         self.server_proc = Process(target=self.server.serve_forever)
         self.server_proc.start()
+        self.is_running = True 
 
     def stopServer(self):
-        self.server.close_all()
-        self.server_proc.terminate()
-        self.server_proc.join()
+        if self.is_running:
+            self.server.close_all()
+            self.server_proc.terminate()
+            self.server_proc.join()
+            del self.server_proc
+            self.server_proc = None 
+            self.is_running = False
